+++ src/accessories/WyzeSwitch.js
@@ -122,19 +122,17 @@ module.exports = class WyzeSwitch extends WyzeAccessory {
       this.plugin.log(
         `[Switch] Target State Set "${this.display_name} (${this.mac})" : "${value}"`
       );
-    try {
-      if (this.single_press_type == SinglePressType.IOT) {
-        await this.plugin.client.wallSwitchIot(this.mac, this.product_model, value ? true : false);
-      } else {
-        await this.plugin.client.wallSwitchPower(this.mac, this.product_model, value ? true : false);
-      }
-      this.switch_power = !!value;
-      this.wallSwitch.getCharacteristic(Characteristic.On).updateValue(this.switch_power);
-    } catch (error) {
+    this.switch_power = !!value;
+    this.wallSwitch.getCharacteristic(Characteristic.On).updateValue(this.switch_power);
+
+    const call = this.single_press_type == SinglePressType.IOT
+      ? this.plugin.client.wallSwitchIot(this.mac, this.product_model, !!value)
+      : this.plugin.client.wallSwitchPower(this.mac, this.product_model, !!value);
+
+    call.catch((error) => {
       this.plugin.log.error?.(
         `[Switch] Failed to set target state for "${this.display_name} (${this.mac})": ${error}`
       );
-      throw error;
-    }
+    });
   }
 };